refactor: 불필요한 클래스 삭제
PlayerIterator 클래스의 기능을 Players로 이동
