+++ src/main/java/blackjack/domain/participant/Players.java
@@ -6,6 +6,7 @@ import java.util.List;
 
 public class Players {
     private final List<Player> players;
+    private int order = 0;
 
     public Players(List<Name> playerNames, HandGenerator handGenerator) {
         checkPlayersEmpty(playerNames);
@@ -18,6 +19,20 @@ public class Players {
         if (playerNames.isEmpty()) {
             throw new InvalidPlayerCountException();
         }
+    }
+
+    public boolean hasNext() {
+        return order < players.size();
+    }
+
+    public void increaseOrder(PlayerAction playerAction) {
+        if (playerAction.equals(PlayerAction.STAND) || !getPlayerAtOrder().canHit()) {
+            order++;
+        }
+    }
+
+    public Player getPlayerAtOrder() {
+        return players.get(order);
     }
 
     public List<Player> getPlayers() {
